Can now get terminal size. Updated cursor.zig to start using u16 instead of u8 variables.

+++ src/cursor.zig
@@ -10,42 +10,42 @@ pub const print = struct {
     }
 
     // move cursor to line x column y
-    pub fn moveTo(x: u8, y: u8) void {
+    pub fn moveTo(x: u16, y: u16) void {
         utils.printEscapeCode("{};{}H", .{x, y});
     }
 
     // move cursor up y lines
-    pub fn moveUp(y: u8) void {
+    pub fn moveUp(y: u16) void {
         utils.printEscapeCode("{}A", .{y});
     }
 
     // move cursor down y lines
-    pub fn moveDown(y: u8) void {
+    pub fn moveDown(y: u16) void {
         utils.printEscapeCode("{}B", .{y});
     }
 
     // move cursor right x lines
-    pub fn moveRight(x: u8) void {
+    pub fn moveRight(x: u16) void {
         utils.printEscapeCode("{}C", .{x});
     }
 
     // move cursor left x lines
-    pub fn moveLeft(x: u8) void {
+    pub fn moveLeft(x: u16) void {
         utils.printEscapeCode("{}D", .{x});
     }
 
     // move cursor to the beginning of next line, y lines down
-    pub fn moveDownStart(y: u8) void {
+    pub fn moveDownStart(y: u16) void {
         utils.printEscapeCode("{}E", .{y});
     }
 
     // move cursor to the beginning of previous line, y lines up
-    pub fn moveUpStart(y: u8) void {
+    pub fn moveUpStart(y: u16) void {
         utils.printEscapeCode("{}F", .{y});
     }
 
     // move cursor to column x
-    pub fn moveToCol(x: u8) void {
+    pub fn moveToCol(x: u16) void {
         utils.printEscapeCode("{}G", .{x});
     }
 };
@@ -58,54 +58,54 @@ pub inline fn reset() []const u8 {
 }
 
 // move cursor to line x column y
-pub inline fn moveTo(x: u8, y: u8) []const u8 {
+pub inline fn moveTo(x: u16, y: u16) []const u8 {
     return utils.returnEscapeCode("{};{}H", .{x, y});
 }
 
 // move cursor up y lines
-pub inline fn moveUp(y: u8) []const u8 {
+pub inline fn moveUp(y: u16) []const u8 {
     return utils.returnEscapeCode("{}A", .{y});
 }
 
 // move cursor down y lines
-pub inline fn moveDown(y: u8) []const u8 {
+pub inline fn moveDown(y: u16) []const u8 {
     return utils.returnEscapeCode("{}B", .{y});
 }
 
 // move cursor right x lines
-pub inline fn moveRight(x: u8) []const u8 {
+pub inline fn moveRight(x: u16) []const u8 {
     return utils.returnEscapeCode("{}C", .{x});
 }
 
 // move cursor left x lines
-pub inline fn moveLeft(x: u8) []const u8 {
+pub inline fn moveLeft(x: u16) []const u8 {
     return utils.returnEscapeCode("{}D", .{x});
 }
 
 // move cursor to the beginning of next line, y lines down
-pub inline fn moveDownStart(y: u8) []const u8 {
+pub inline fn moveDownStart(y: u16) []const u8 {
     return utils.returnEscapeCode("{}E", .{y});
 }
 
 // move cursor to the beginning of previous line, y lines up
-pub inline fn moveUpStart(y: u8) []const u8 {
+pub inline fn moveUpStart(y: u16) []const u8 {
     return utils.returnEscapeCode("{}F", .{y});
 }
 
 // move cursor to column x
-pub inline fn moveToCol(x: u8) []const u8 {
+pub inline fn moveToCol(x: u16) []const u8 {
     return utils.returnEscapeCode("{}G", .{x});
 }
 
 // get cursor position in the format [row; column]
 // requires raw mode to be enabled
-pub fn getPosition() ![2]u8 {
+pub fn getPosition() ![2]u16 {
     const stdin = std.io.getStdIn().reader();
     const stdout = std.io.getStdOut().writer();
 
     try stdout.writeAll(utils.returnEscapeCode("6n", .{}));
 
-    var pos = [2]u8{0,0};
+    var pos = [2]u16{0,0};
     var buffer: [32]u8 = undefined;
     var index: usize = 0;
     var pos_i: u8 = 0;
